Add environment variable to control mongodb connection string
This change is backwards-compatible.

When running the backend, we can set environment variables to
make sure that the mongodb connection string goes to the right
place. If no environment variables are set, then the connection
string defaults to what it was before: mongodb://mongo/guesstimoji.

I added this change because my computer ran mongodb on
mongodb://localhost/guesstimoji, so the backend could not connect
correctly.