Replace named URL hosts with localhost
This change may be bad, in which case we can revert it.

I made it because my machine does not name the other hosts, so the
frontend cannot correctly connect to the backend. I changed them
all to localhost, which runs correctly.

This might cause problems when developing on other machines or
during deployment. If so, then revert this commit.