Remove old logging statements and commented code
These are useful to debug, but they're used for parts that we know
work correctly. It's better to remove them so they don't distract
us from more important debugging statements. Note that this only
applies to the frontend, and I kept console.log statements that
are important, like the statement that says that the user is
offline.

I also remove the commented-out code because, by now, we know that
we won't reuse it.