Restart game on button press
This exposes a button that restarts the entire game. Note that this
does not change the board that the players are using or the emojis
that they picked. Instead, it resets the game state to empty.

The next step is to make it so that the picked emojis are both in the
game state, so reset the game state lets the users pick a different
emoji. I'll probably make it so that the board does not change because
I don't want to implement board re-selection.

Also, I want to make it so that the restart button only appears at a
game over. Otherwise, people might abuse the button.