Do not let players guess until everyone has picked
If player 2 has picked an emoji but player 1 has not, then we should not
let player 2 make guesses or send messages. However, the code allowed
player 2 to do so.

This commit adds a room state variable which tracks whether all the
players are ready to play the game. When the room starts, the variable
is set to false. After a player sends their pick to the server, the
server checks to see if both players have made picks. If so, the server
sends an action that tells the clients that everyone is ready. The
clients then update their own states, and the messaging box is allowed
to display.

This resolves Issue #9 on the main GitHub repository.