Close socket on component end
This is a small error. `useEffect` should return a function that closes
the socket, but we accidentally closed the socket immediately. This
meant that the component was not listening for `server:players/reset`
messages, so the server could not ask the client to reset the game.
This meant that, when a player tried to reset the game, it reset on
their machine but not on the other player's machine.

This commit fixes that.